feat: DownloadLibraries for downloading in afterEvaluated
Took 7 minutes
